Move App.vue into view directory
