+++ src/renderer/view/App.vue
@@ -85,27 +85,27 @@ import { AppState, ResearchState } from "@/common/control/state.js";
 import AnalysisDialog from "@/renderer/view/dialog/AnalysisDialog.vue";
 import CSAGameReadyDialog from "@/renderer/view/dialog/CSAGameReadyDialog.vue";
 import { CSAGameState } from "@/renderer/store/csa";
-import { useAppSettings } from "./store/settings";
+import { useAppSettings } from "@/renderer/store/settings";
 import { BackgroundImageType } from "@/common/settings/app";
-import MateSearchDialog from "./view/dialog/MateSearchDialog.vue";
-import PVPreviewDialog from "./view/dialog/PVPreviewDialog.vue";
-import RecordFileHistoryDialog from "./view/dialog/RecordFileHistoryDialog.vue";
-import BatchConversionDialog from "./view/dialog/BatchConversionDialog.vue";
-import LaunchUSIEngineDialog from "./view/dialog/LaunchUSIEngineDialog.vue";
-import ConnectToCSAServerDialog from "./view/dialog/ConnectToCSAServerDialog.vue";
-import PieceSetChangeDialog from "./view/dialog/PieceSetChangeDialog.vue";
-import LoadRemoteFileDialog from "./view/dialog/LoadRemoteFileDialog.vue";
-import ShareDialog from "./view/dialog/ShareDialog.vue";
-import AddBookMovesDialog from "./view/dialog/AddBookMovesDialog.vue";
-import { useBusyState } from "./store/busy";
-import { useMessageStore } from "./store/message";
-import { useErrorStore } from "./store/error";
-import { useConfirmationStore } from "./store/confirm";
-import CustomLayout from "./view/main/CustomLayout.vue";
-import MobileLayout from "./view/main/MobileLayout.vue";
-import api, { isMobileWebApp, isNative } from "./ipc/api";
-import { openCopyright } from "./helpers/copyright";
-import { installHotKeyForMainWindow } from "./devices/hotkey";
+import MateSearchDialog from "./dialog/MateSearchDialog.vue";
+import PVPreviewDialog from "./dialog/PVPreviewDialog.vue";
+import RecordFileHistoryDialog from "./dialog/RecordFileHistoryDialog.vue";
+import BatchConversionDialog from "./dialog/BatchConversionDialog.vue";
+import LaunchUSIEngineDialog from "./dialog/LaunchUSIEngineDialog.vue";
+import ConnectToCSAServerDialog from "./dialog/ConnectToCSAServerDialog.vue";
+import PieceSetChangeDialog from "./dialog/PieceSetChangeDialog.vue";
+import LoadRemoteFileDialog from "./dialog/LoadRemoteFileDialog.vue";
+import ShareDialog from "./dialog/ShareDialog.vue";
+import AddBookMovesDialog from "./dialog/AddBookMovesDialog.vue";
+import { useBusyState } from "@/renderer/store/busy";
+import { useMessageStore } from "@/renderer/store/message";
+import { useErrorStore } from "@/renderer/store/error";
+import { useConfirmationStore } from "@/renderer/store/confirm";
+import CustomLayout from "./main/CustomLayout.vue";
+import MobileLayout from "./main/MobileLayout.vue";
+import api, { isMobileWebApp, isNative } from "@/renderer/ipc/api";
+import { openCopyright } from "@/renderer/helpers/copyright";
+import { installHotKeyForMainWindow } from "@/renderer/devices/hotkey";
 import { DialogPosition } from "@/common/settings/layout";
 
 const clipboard = ref();
